Create SSH connection with Paramiko
if you want to write a program that will log into a device and check for a condition periodically you can use Paramko. Paramiko is a low-level Python package which allows you to connect to a device via SSH. Paramiko is considered low level because it does not know to wait for a specified amount of time before an action is complete. If you try to use paramiko to, for example, take a snapshot of a network device's running config, depending on the device and the size of the config it might take several seconds for the device to generate the config and pass it to the python Paramiko package. In order to get this information accurately one should use the time.sleep function in order to create delays. 

Paramiko only works with SSH. Because it works with SSH it will first check to see if the host key provided by device that it's attempting to connect to is known. If it is not known, the connection will not work. Therefore, unless the device you are running the Python script on already has the host key for the device that yo are connecting to you will have to let Paramiko know to proceed with unknown host keys. 

The first section of the code only imports the classes of paramiko that we will be working on. AutoAddPolicy is to allow for unknown host keys and SSHClient is the handle that we will use to connect to the devices that we will be interacting with.